Correctly typecheck lists in records
As previously stated, it could be derived from the typing rules that a
record of lists is a subtype of a record of single elements:
e.g. that (x: List Num, y: Bool) ⊆ (x: Num, y: Bool). Though this
semantics makes a certain amount of sense given the existing behaviour
of lists wrt typechecking, it is not the semantics I prefer for this
language. Additionally, it poses problems with field access:
continuing the example, from (x: List Num, y: Bool) ⊆ (x: T) we could
infer that T ∼ Num unlifted once, which requires further special
handling in other parts of the typechecker¹. Accordingly the types of
record elements must now be equivalent, rather than just subtypes.

______________
¹ Though this is arguably consistent with the existing behaviour of
functions, so this decision may be reverted in the future.