Bugfix: substitute new type variables for tables as they are found
Previously tables were typechecked by generating type variables a,b,…
for each column, typechecking each column (which may give constraints
against other variables), then unifying each column and applying the
resulting substitution. However this fails when operations such as
unlifting are required, since the exact identities of the type
variables are unknown at the time of column typechecking. By
substituting the type variables for their known values as each column
is checked, this problem can be avoided.

Of course this new approach requires the columns to be typechecked in
the right order. Topological sorting is necessary to get this working
right; for now this commit uses the simple but invalid method of
checking each column in order.