Evaluate all cells in same environment
Previously each cell was evaluated in a different environment, with
new environments being set up and torn down on each call to the
typechecker or evaluator. This made parts of the code somewhat
difficult to comprehend. By running the whole computation in one
global environment, the code is made easier to
understand. Additionally this allows for namespacing and more granular
scoping rules to be implemented in the future if necessary.