Merge table definition with recursive records
Conceptually, tables are simply a type of record. To reflect this in
the formula language, most table functionality has been transferred
over to records, in the form of 'recursive' record definitions where
fields can refer to other fields (an idea also present in the Nix
expression language). Tables have been retained as a separate data
type to provide additional type-level guarantees in areas such as indexing.